+++ spec/integrations/cli/cli_unknown_spec.rb
@@ -1,8 +1,10 @@
 # frozen_string_literal: true
 
-# Karafka CLI should work and should just run help without command
+# Karafka CLI should work and should fail with unknown command error
 
 failed = false
+
+ARGV[0] = 'unknown'
 
 begin
   Karafka::Cli.start
@@ -10,4 +12,4 @@ rescue Karafka::Errors::UnrecognizedCommandError
   failed = true
 end
 
-assert !failed
+assert failed