+++ CHANGELOG.md
@@ -32,6 +32,9 @@
 - [Enhancement] Allow for usage of `latest` and `earliest` in the `Karafka::Pro::Iterator`.
 - [Enhancement] Failures during `topics migrate` (and other subcommands) don't show what topic failed, and why it's invalid.
 - [Enhancement] Apply changes to topics configuration in atomic independent requests when using Declarative Topics.
+- [Enhancement] Execute the help CLI command when no command provided (similar to Rails) to improve DX.
+- [Enhancement] Remove backtrace from the CLI error for incorrect commands (similar to Rails) to improve DX.
+- [Enhancement] Provide `karafka topics help` sub-help due to nesting of Declarative Topics actions.
 - [Refactor] Introduce a `bin/verify_kafka_warnings` script to clean Kafka from temporary test-suite topics.
 - [Refactor] Introduce a `bin/verify_topics_naming` script to ensure proper test topics naming convention.
 - [Refactor] Make sure all temporary topics have a `it-` prefix in their name.